+++ .eslintrc.js
@@ -1,0 +1,9 @@
+"use strict";
+
+module.exports = {
+  "extends": [
+      "./packages/eslint-config-spt/index.js",
+      "./packages/eslint-config-spt-node/index.js"
+  ],
+  "root": true
+};