+++ packages/eslint-config-spt-react/airbnb.js
@@ -1,0 +1,6 @@
+"use strict";
+
+module.exports = {
+  extends: ["./rules/react", "./rules/react-a11y"].map(require.resolve),
+  rules: {}
+};