+++ packages/eslint-config-spt-react/index.js
@@ -1,0 +1,5 @@
+"use strict";
+
+module.exports = {
+  extends: ["./react.js", "./airbnb.js", "./prettier.js"].map(require.resolve)
+};