+++ packages/eslint-config-spt-react/react.js
@@ -1,0 +1,11 @@
+"use strict";
+
+module.exports = {
+  plugins: ["react"],
+  parserOptions: {
+    ecmaFeatures: {
+      jsx: true
+    }
+  },
+  rules: {}
+};